+++ packages/client/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.44.0
+* Add `isStale`, sugar for `markedForUpdate || updating`
+* Node flags `isStale`, `markedForUpdate`, and `updating` are now properly set at the group level as whether any of their children have them set
+* Deprecated undocumented internal `flat` tree, will remove in next version
+* Remove unused `default` type support
+* `watchNode` now supports nested paths (e.g. `context.something`) and empty keys array
+* Actions now recieve a full `TreeInstance` instead of just action props
+
 # 2.43.0
 * Add `watchNode`